dataをpostできるように変更
