+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "@toast-ui/react-image-editor": "^1.0.1",
     "@types/jest": "^24.0.0",
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.0",
@@ -22,12 +23,12 @@
     "react-dom": "^16.12.0",
     "react-redux": "^7.1.3",
     "react-scripts": "3.3.0",
+    "reapop": "^2.1.0",
     "redux-form": "^8.2.6",
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.1.3",
     "typescript": "~3.7.2",
-    "typescript-fsa": "^3.0.0",
-    "@toast-ui/react-image-editor": "^1.0.1"
+    "typescript-fsa": "^3.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
