+++ src/App.tsx
@@ -5,7 +5,7 @@ import SendForm from "./components/sendForm"
 import {Dispatch} from "redux";
 import {Action} from "typescript-fsa";
 import {SetImageDataActionCreator} from "./modules/Image";
-import ImageEditor from "./components/ImageEditor";
+import ImageEditor from "./components/ImageEditorComponent";
 
 interface Props {
     loadData(): void
@@ -17,7 +17,7 @@ const App = (props: Props) => {
     })
   return (
     <div>
-        <SendForm />
+        <SendForm onSubmit={(data)=> console.log(data)} />
         <ImageEditor/>
     </div>
