+++ src/store.tsx
@@ -4,14 +4,14 @@ import reduxLogger from "redux-logger";
 import createSagaMiddleware from "redux-saga";
 import rootSaga from "./RootSaga";
 
-
 export const sagaMiddleware = createSagaMiddleware();
 
 
 export const buildStore = (
     configureStore({
         reducer: rootReducer,
-        middleware: [sagaMiddleware, reduxLogger]
+        middleware: [sagaMiddleware, reduxLogger],
+
     })
 )
